feat: add vibrator binding
