+++ Cargo.toml
@@ -21,6 +21,8 @@ ohos-native-buffer-binding          = { version = "0.1.0", path = "crates/native
 ohos-udmf-binding                   = { version = "0.0.1", path = "crates/udmf" }
 ohos-pasteboard-binding             = { version = "0.0.1", path = "crates/pasteboard" }
 ohos-web-binding                    = { version = "0.1.0", path = "crates/web" }
+ohos-sensor-binding                 = { version = "0.1.0", path = "crates/sensor" }
+ohos-vibrator-binding               = { version = "0.1.0", path = "crates/vibrator" }
 
 ohos-bundle-sys                 = { version = "0.0.2", path = "sys/ohos-bundle-sys" }
 ohos-init-sys                   = { version = "0.0.2", path = "sys/ohos-init-sys" }
@@ -43,7 +45,7 @@ ohos-pasteboard-sys             = { version = "0.0.1", path = "sys/ohos-pasteboa
 ohos-image-sys                  = { version = "0.0.1", path = "sys/ohos-image-sys" }
 ohos-image-native-sys           = { version = "0.0.1", path = "sys/ohos-image-native-sys" }
 ohos-web-sys                    = { version = "0.0.1", path = "sys/ohos-web-sys" }
-ohos-sensors-sys                = { version = "0.1.0", path = "sys/ohos-sensors-sys" }
+ohos-sensor-sys                = { version = "0.1.0", path = "sys/ohos-sensor-sys" }
 ohos-vibrator-sys               = { version = "0.1.0", path = "sys/ohos-vibrator-sys" }
 
 ohos_enum_macro = { version = "0.0.2", path = "crates/enum_macro" }
