+++ crates/vibrator/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "ohos-vibrator-binding"
+version = "0.1.0"
+edition = "2021"
+license     = "MIT OR Apache-2.0"
+description = "OpenHarmony's vibrator binding for rust"
+
+[dependencies]
+ohos-vibrator-sys = { workspace = true }
+ohos_enum_macro  = { workspace = true }
