+++ sys/ohos-sensor-sys/Cargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "ohos-sensor-sys"
+version = "0.1.0"
+edition = "2021"
+license     = "MIT OR Apache-2.0"
+description = "OpenHarmony's sensor sys binding for rust"
+
+[dependencies]
