--- sys/ohos-sensors-sys/Cargo.toml
@@ -1,8 +1,0 @@
-[package]
-name = "ohos-sensors-sys"
-version = "0.1.0"
-edition = "2021"
-license     = "MIT OR Apache-2.0"
-description = "OpenHarmony's sensors binding for rust"
-
-[dependencies]
