+++ sys/ohos-vibrator-sys/Cargo.toml
@@ -3,6 +3,6 @@ name = "ohos-vibrator-sys"
 version = "0.1.0"
 edition = "2021"
 license     = "MIT OR Apache-2.0"
-description = "OpenHarmony's vibrator binding for rust"
+description = "OpenHarmony's vibrator sys binding for rust"
 
 [dependencies]
